Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
 
 int main() {
-double a;
-scanf("%lf",&a);
-// l is use for double
-printf("You entered: %lf",a);
+float a;
+scanf("%.4f",&a);
+
+printf("You entered: %.4f",a);
 
 }
